adding validator

+++ controllers/contactsController.js
@@ -3,26 +3,32 @@ const objectId = require('mongodb').ObjectId;
 
 const getAll = async (req, res) => {
     //#swagger.tags = ['Contacts']
-    mongodb.getDatabase().db().collection('contacts').find().toArray((err, contacts) => {
-        if (err) {
-            res.status(400).json({message: err});
-        }
+    const result = await mongodb.getDatabase().db().collection('contacts').find();
+    result.toArray().then((contacts) => {
         res.setHeader('Content-Type', 'application/json');
         res.status(200).json(contacts);
+        }).catch((err) => {
+        res.status(500).json({ message: 'Failed to retrieve contacts.', error: err });
     });
 };
 
 const getSingle = async (req, res) => {
+    // validator
+    if (!objectId.isValid(req.params.id)) {
+        res.status(400).json('Invalid contact ID format');
+        return;
+    }
+
     //#swagger.tags = ['Contacts']
     const userId = new objectId(req.params.id);
 
-    mongodb.getDatabase().db().collection('contacts').find({ _id: userId }).toArray((err, result) => {
-        if (err) {
-            res.status(400).json({message: err});
-        }
+    const result = await mongodb.getDatabase().db().collection('contacts').find({ _id: userId });
+    result.toArray().then((contacts) => {
         res.setHeader('Content-Type', 'application/json');
-        res.status(200).json(result[0]);
-    });
+        res.status(200).json(contacts[0]);
+    }).catch((err) => { 
+        res.status(500).json({ message: 'Failed to retrieve contact.', error: err });
+    });    
 };
 
 const createContact = async (req, res) => {
@@ -43,6 +49,12 @@ const createContact = async (req, res) => {
 };
 
 const updateContact = async (req, res) => {
+    // validator
+    if (!objectId.isValid(req.params.id)) {
+        res.status(400).json('Invalid contact ID format');
+        return;
+    }   
+
     //#swagger.tags = ['Contacts']
     const userId = new objectId(req.params.id);
     const contact = {
@@ -61,6 +73,12 @@ const updateContact = async (req, res) => {
 };
 
 const deleteContact = async (req, res) => {
+    // validator
+    if (!objectId.isValid(req.params.id)) {
+        res.status(400).json('Invalid contact ID format');
+        return;
+    }
+
     //#swagger.tags = ['Contacts']
     const userId = new objectId(req.params.id);
     const result = await mongodb.getDatabase().db().collection('contacts').deleteOne({ _id: userId });
@@ -77,4 +95,4 @@ module.exports = {
     createContact,
     updateContact,
     deleteContact
-};
+};
